Moved classes in new namespaces
